src/holopy/quantum/measurement.py
"""
Holographic measurement system with quantum state tomography.
"""
from typing import Dict, List, Tuple, Optional
import numpy as np
from scipy.optimize import minimize
import logging
from dataclasses import dataclass
from .error_correction import HolographicStabilizer
from ..config.constants import (
    INFORMATION_GENERATION_RATE,
    PLANCK_CONSTANT,
    SPEED_OF_LIGHT
)

logger = logging.getLogger(__name__)

@dataclass
class MeasurementResult:
    """Container for measurement results and uncertainty."""
    expectation_value: float
    uncertainty: float
    basis_state: str
    collapse_fidelity: float
    information_gain: float

@dataclass
class TomographyResult:
    """Results from quantum state tomography."""
    density_matrix: np.ndarray
    fidelity: float
    purity: float
    entropy: float
    confidence: float

class HolographicMeasurement:
    """Implements holographic quantum measurements and tomography."""
    
    def __init__(
        self,
        n_qubits: int,
        measurement_bases: Optional[List[str]] = None,
        error_correction: Optional[HolographicStabilizer] = None
    ):
        """
        Initialize holographic measurement system.
        
        Args:
            n_qubits: Number of qubits
            measurement_bases: Optional list of measurement bases
            error_correction: Optional error correction system
        """
        self.n_qubits = n_qubits
        self.error_correction = error_correction
        self.measurement_bases = measurement_bases or ['Z']  # Default to Z-basis
        
        # Initialize measurement operators
        self._initialize_operators()
        
        logger.info(f"Initialized HolographicMeasurement for {n_qubits} qubits")
    
    def _initialize_operators(self) -> None:
        """Initialize measurement operators for each basis."""
        try:
            self.operators = {}
            
            # Pauli operators
            sigma_x = np.array([[0, 1], [1, 0]], dtype=complex)
            sigma_y = np.array([[0, -1j], [1j, 0]], dtype=complex)
            sigma_z = np.array([[1, 0], [0, -1]], dtype=complex)
            
            # Create measurement operators for each basis
            for basis in self.measurement_bases:
                if basis == 'X':
                    self.operators[basis] = sigma_x
                elif basis == 'Y':
                    self.operators[basis] = sigma_y
                elif basis == 'Z':
                    self.operators[basis] = sigma_z
                else:
                    raise ValueError(f"Unknown measurement basis: {basis}")
            
            logger.debug(f"Initialized operators for bases: {self.measurement_bases}")
            
        except Exception as e:
            logger.error(f"Operator initialization failed: {str(e)}")
            raise
    
    def measure_state(
        self,
        state: np.ndarray,
        basis: str,
        qubit: int
    ) -> MeasurementResult:
        """
        Perform quantum measurement with holographic constraints.
        
        Args:
            state: Quantum state vector
            basis: Measurement basis
            qubit: Target qubit
            
        Returns:
            MeasurementResult containing measurement outcome
        """
        try:
            # Verify basis
            if basis not in self.operators:
                raise ValueError(f"Invalid measurement basis: {basis}")
            
            # Get measurement operator
            operator = self._expand_operator(self.operators[basis], qubit)
            
            # Calculate expectation value with holographic noise
            expectation = np.real(np.vdot(state, operator @ state))
            noise = np.random.normal(0, INFORMATION_GENERATION_RATE)
            measured_value = expectation + noise
            
            # Calculate measurement uncertainty
            uncertainty = self._calculate_uncertainty(state, operator)
            
            # Project state according to measurement
            collapsed_state, fidelity = self._project_state(
                state,
                operator,
                measured_value
            )
            
            # Calculate information gain
            information_gain = self._calculate_information_gain(
                state,
                collapsed_state
            )
            
            result = MeasurementResult(
                expectation_value=measured_value,
                uncertainty=uncertainty,
                basis_state=f"{basis}{qubit}",
                collapse_fidelity=fidelity,
                information_gain=information_gain
            )
            
            logger.debug(
                f"Measured qubit {qubit} in {basis}-basis: "
                f"value={measured_value:.4f} ± {uncertainty:.4f}"
            )
            
            return result
            
        except Exception as e:
            logger.error(f"Measurement failed: {str(e)}")
            raise
    
    def perform_tomography(
        self,
        state: np.ndarray,
        n_measurements: int = 1000
    ) -> TomographyResult:
        """
        Perform quantum state tomography with holographic constraints.
        
        Args:
            state: Quantum state to reconstruct
            n_measurements: Number of measurements per basis
            
        Returns:
            TomographyResult containing reconstructed state
        """
        try:
            # Collect measurements in each basis
            measurement_data = []
            
            for basis in self.measurement_bases:
                for qubit in range(self.n_qubits):
                    basis_measurements = []
                    for _ in range(n_measurements):
                        result = self.measure_state(state.copy(), basis, qubit)
                        basis_measurements.append(result.expectation_value)
                    measurement_data.append((basis, qubit, basis_measurements))
            
            # Reconstruct density matrix
            rho = self._reconstruct_density_matrix(measurement_data)
            
            # Calculate tomography metrics
            fidelity = self._calculate_state_fidelity(state, rho)
            purity = np.real(np.trace(rho @ rho))
            entropy = -np.real(np.trace(rho @ np.log2(rho + 1e-10)))
            confidence = self._calculate_tomography_confidence(
                measurement_data,
                rho
            )
            
            result = TomographyResult(
                density_matrix=rho,
                fidelity=fidelity,
                purity=purity,
                entropy=entropy,
                confidence=confidence
            )
            
            logger.info(
                f"Completed tomography with {n_measurements} measurements "
                f"per basis, fidelity: {fidelity:.4f}"
            )
            
            return result
            
        except Exception as e:
            logger.error(f"Tomography failed: {str(e)}")
            raise
    
    def _expand_operator(
        self,
        operator: np.ndarray,
        target_qubit: int
    ) -> np.ndarray:
        """Expand single-qubit operator to full system size."""
        try:
            expanded = np.eye(1, dtype=complex)
            
            for i in range(self.n_qubits):
                if i == target_qubit:
                    expanded = np.kron(expanded, operator)
                else:
                    expanded = np.kron(expanded, np.eye(2))
            
            return expanded
            
        except Exception as e:
            logger.error(f"Operator expansion failed: {str(e)}")
            raise
    
    def _calculate_uncertainty(
        self,
        state: np.ndarray,
        operator: np.ndarray
    ) -> float:
        """Calculate measurement uncertainty with holographic effects."""
        try:
            # Calculate quantum uncertainty
            expectation_sq = np.real(np.vdot(state, operator @ operator @ state))
            expectation = np.real(np.vdot(state, operator @ state))
            variance = expectation_sq - expectation**2
            
            # Add holographic noise
            total_uncertainty = np.sqrt(
                variance + INFORMATION_GENERATION_RATE**2
            )
            
            return total_uncertainty
            
        except Exception as e:
            logger.error(f"Uncertainty calculation failed: {str(e)}")
            raise 